requirements.txt
Flask==2.3.2
gunicorn==23.0.0
celery==5.2.7
kombu==5.2.4
redis==3.5.3
pdfminer.six==20221105
nltk==3.8.1
langdetect==1.0.9
openai==0.27.10
tiktoken==0.4.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
